修改管理命令 user_day_net_io
